Add entry collection decorator for Archive
